27 - Remove Element
In this problem, we need to remove all instances of a given value and
return the new length of the array. In my approach, if the value is
found, I swap it with the leftmost available space for it.

Time Complexity: O(n)
Space Complexity: O(1)

Notes:
I could have avoided swapping by simply overwriting the value at the
leftmost index with the current one, but I decided not to delete info.