26 - Remove Duplicates from Sorted Array
Similar to 27, the point of this is to remove elements except for the
first instance of each unique element. Originally I used a set to keep
track of the unique elements, but I realized that since the values are
sorted in non-decreasing order, I can just keep track of the amount of
unique elements by checking if the next element is greater than the
previous. Since we don't care about the elements after the unique ones,
it doesn't matter what they are and can even include duplicates.

Time Complexity: O(n)
Space Complexity: O(1)

Notes:
In a real program, it doesn't make sense to have bad values remain in
the original array and the should be resized to account for this.